update to work with pico
